+++ src/main/java/frc/robot/subsystems/sinigle_motor/SingleMotorIoReplay.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.general;
+package frc.robot.subsystems.sinigle_motor;
 
 public class SingleMotorIoReplay implements SingleMotorIo {
