+++ src/main/java/frc/robot/subsystems/sinigle_motor/SingleMotorIoSparkMax.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.general;
+package frc.robot.subsystems.sinigle_motor;
 
 import com.revrobotics.CANSparkBase.IdleMode;
 import com.revrobotics.CANSparkLowLevel.MotorType;
