+++ src/main/java/frc/robot/subsystems/swerve_module/IndexedSwerveModule.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.swerve;
+package frc.robot.subsystems.swerve_module;
 
 import org.littletonrobotics.junction.Logger;
