+++ src/main/java/frc/robot/subsystems/swerve_module/SwerveModuleIoTalonFx.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.swerve;
+package frc.robot.subsystems.swerve_module;
 
 import com.ctre.phoenix6.BaseStatusSignal;
 import com.ctre.phoenix6.StatusSignal;
